+++ ft_strlcpy.c
@@ -13,8 +13,8 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	size_t	i;
-	int		slen;
+	size_t		i;
+	size_t		slen;
 
 	i = 0;
 	slen = ft_strlen((char *)src);
